+++ small_text/integrations/transformers/classifiers/__init__.py
@@ -1,3 +1,10 @@
 from small_text.integrations.transformers.classifiers.base import ModelLoadingStrategy
 from small_text.integrations.transformers.classifiers.classification import TransformerBasedClassification
 from small_text.integrations.transformers.classifiers.factories import TransformerBasedClassificationFactory
+
+
+__all__ = [
+    'ModelLoadingStrategy',
+    'TransformerBasedClassification',
+    'TransformerBasedClassificationFactory'
+]